#1054 Replace name-based identifier with qualifiers in dependency declarations
