+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/inject/DependencyDeclarationContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2023 the original author or authors.
+ * Copyright 2019-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,15 @@
 
 package org.dockbox.hartshorn.inject;
 
+import org.dockbox.hartshorn.component.CompositeQualifier;
 import org.dockbox.hartshorn.util.introspect.view.TypeView;
 
 public interface DependencyDeclarationContext<T> {
 
     TypeView<T> type();
 
+    CompositeQualifier qualifier();
+
+    @Deprecated(since = "0.6.0", forRemoval = true)
     String id();
 }
