+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/inject/PostProcessorDependencyDeclarationContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2023 the original author or authors.
+ * Copyright 2019-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,14 +16,23 @@
 
 package org.dockbox.hartshorn.inject;
 
+import org.dockbox.hartshorn.component.CompositeQualifier;
 import org.dockbox.hartshorn.component.processing.ComponentPostProcessor;
 import org.dockbox.hartshorn.util.introspect.view.TypeView;
 
 public record PostProcessorDependencyDeclarationContext<T extends ComponentPostProcessor>(
-        TypeView<T> type) implements DependencyDeclarationContext<T> {
+        TypeView<T> type
+) implements DependencyDeclarationContext<T> {
+
+    @Override
+    public CompositeQualifier qualifier() {
+        // Post processors are not qualified in any way, so we can return an empty qualifier here
+        return new CompositeQualifier();
+    }
 
     @Override
     public String id() {
+        // Post processors are not qualified in any way, so we can return null here
         return null;
     }
 }
